+++ molgenis-core-ui/src/main/java/org/molgenis/ui/admin/user/UserAccountController.java
@@ -5,9 +5,9 @@ import org.apache.commons.lang3.StringUtils;
 import org.molgenis.auth.User;
 import org.molgenis.data.i18n.LanguageService;
 import org.molgenis.data.settings.AppSettings;
-import org.molgenis.security.twofactor.RecoveryCode;
-import org.molgenis.security.twofactor.RecoveryService;
-import org.molgenis.security.twofactor.TwoFactorAuthenticationService;
+import org.molgenis.security.twofactor.meta.RecoveryCode;
+import org.molgenis.security.twofactor.service.RecoveryService;
+import org.molgenis.security.twofactor.service.TwoFactorAuthenticationService;
 import org.molgenis.security.user.MolgenisUserException;
 import org.molgenis.security.user.UserAccountService;
 import org.molgenis.ui.MolgenisPluginController;
