+++ molgenis-security/src/main/java/org/molgenis/security/MolgenisWebAppSecurityConfig.java
@@ -21,6 +21,10 @@ import org.molgenis.security.token.TokenAuthenticationFilter;
 import org.molgenis.security.token.TokenAuthenticationProvider;
 import org.molgenis.security.token.TokenGenerator;
 import org.molgenis.security.twofactor.*;
+import org.molgenis.security.twofactor.auth.*;
+import org.molgenis.security.twofactor.service.OTPService;
+import org.molgenis.security.twofactor.service.RecoveryService;
+import org.molgenis.security.twofactor.service.TwoFactorAuthenticationService;
 import org.molgenis.security.user.MolgenisUserDetailsChecker;
 import org.molgenis.security.user.UserAccountService;
 import org.molgenis.security.user.UserDetailsService;
