+++ molgenis-security/src/main/java/org/molgenis/security/twofactor/auth/TwoFactorAuthenticationProviderImpl.java
@@ -1,5 +1,8 @@
-package org.molgenis.security.twofactor;
+package org.molgenis.security.twofactor.auth;
 
+import org.molgenis.security.twofactor.service.OTPService;
+import org.molgenis.security.twofactor.service.RecoveryService;
+import org.molgenis.security.twofactor.service.TwoFactorAuthenticationService;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
