+++ molgenis-security/src/main/java/org/molgenis/security/twofactor/meta/RecoveryCode.java
@@ -1,4 +1,4 @@
-package org.molgenis.security.twofactor;
+package org.molgenis.security.twofactor.meta;
 
 import org.molgenis.data.Entity;
 import org.molgenis.data.meta.model.EntityType;
