+++ molgenis-security/src/test/java/org/molgenis/security/twofactor/auth/TwoFactorAuthenticationProviderTest.java
@@ -1,5 +1,6 @@
-package org.molgenis.security.twofactor;
+package org.molgenis.security.twofactor.auth;
 
+import org.molgenis.security.twofactor.service.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
